BugBounty/requirements.txt
cryptography>=41.0.0
pyotp>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
twilio>=8.5.0
jwt>=1.3.1
PyJWT>=2.8.0 